requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
python-dotenv==1.1.0
pydantic==2.11.7
pydantic-settings==2.9.1
Jinja2==3.1.6
python-multipart==0.0.20
pytz==2025.2
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
asyncpg==0.30.0
alembic==1.16.2
cryptography==45.0.4
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.3.0
requests==2.32.4
itsdangerous==2.2.0
starlette==0.46.2

# ML Signal Service Dependencies
pandas==2.3.0
numpy==2.2.6
xgboost==2.1.3
scikit-learn==1.5.2

